+++ src/test/kotlin/org/koin/test/example/MockingInjectionTest.kt
@@ -1,9 +1,9 @@
 package org.koin.test.example
 
 import org.junit.Test
-import org.koin.test.ServiceA
-import org.koin.test.ServiceB
-import org.koin.test.ServiceC
+import org.koin.test.koin.example.ServiceA
+import org.koin.test.koin.example.ServiceB
+import org.koin.test.koin.example.ServiceC
 import org.mockito.Mockito.*
 
 /**