+++ src/test/kotlin/org/koin/test/koin/KoinTest.kt
@@ -4,12 +4,8 @@ import org.junit.Assert
 import org.junit.Assert.*
 import org.junit.Test
 import org.koin.Koin
-import org.koin.test.ServiceA
-import org.koin.test.ServiceB
-import org.koin.test.ServiceC
-import org.koin.test.koin.example.SampleModuleA
-import org.koin.test.koin.example.SampleModuleAC
-import org.koin.test.koin.example.SampleModuleB
+import org.koin.error.NoBeanDefFoundException
+import org.koin.test.koin.example.*
 import org.mockito.Mockito
 import org.mockito.Mockito.times
 
@@ -19,41 +15,89 @@ import org.mockito.Mockito.times
 class KoinTest {
 
     @Test
-    fun `simple load and retrieve instance`() {
-        val ctx = Koin().build(SampleModuleB::class)
+    fun `functional provide component`() {
+        val ctx = Koin().build()
 
-        val serviceB = ctx.get<ServiceB>()
+        val serviceB: ServiceB = Mockito.mock(ServiceB::class.java)
+        Mockito.`when`(serviceB.doSomething()).then {
+            println("done B Mock")
+        }
 
+        ctx.provide { serviceB }
+        ctx.provide({ ServiceA(ctx.get()) })
+
+        assertEquals(2, ctx.beanRegistry.definitions.size)
+        assertEquals(0, ctx.beanRegistry.instanceFactory.instances.size)
+
+        val serviceA_1 = ctx.get<ServiceA>()
+        serviceA_1.doSomethingWithB()
+
+        val serviceA_2 = ctx.get<ServiceA>()
+        serviceA_2.doSomethingWithB()
+
+        assertEquals(2, ctx.beanRegistry.definitions.size)
+        assertEquals(2, ctx.beanRegistry.instanceFactory.instances.size)
+
+        assertEquals(serviceA_1, serviceA_2)
+        Mockito.verify(serviceB, times(2)).doSomething()
+    }
+
+    @Test
+    fun `class providing component`() {
+        val ctx = Koin().build()
+
+        ctx.provide(ServiceB::class)
+        ctx.provide(ServiceA::class)
+
+        assertEquals(2, ctx.beanRegistry.definitions.size)
+        assertEquals(0, ctx.beanRegistry.instanceFactory.instances.size)
+
+        assertNotNull(ctx.get<ServiceB>())
+        assertNotNull(ctx.get<ServiceA>())
+
+        assertEquals(2, ctx.beanRegistry.definitions.size)
+        assertEquals(2, ctx.beanRegistry.instanceFactory.instances.size)
+
+    }
+
+    @Test
+    fun `missing bean component`() {
+        val ctx = Koin().build()
+
+        ctx.provide(ServiceB::class)
+
+        assertNull(ctx.getOrNull<ServiceA>())
         assertEquals(1, ctx.beanRegistry.definitions.size)
-        assertEquals(1, ctx.beanRegistry.instanceFactory.instances.size)
-
-        val serviceA = ctx.getOrNull<ServiceA>()
-
-        assertNotNull(serviceB)
-        assertNull(serviceA)
+        assertEquals(0, ctx.beanRegistry.instanceFactory.instances.size)
     }
 
     @Test
-    fun `load mulitple definition and retrieve them`() {
-        val ctx = Koin().build(SampleModuleAC::class)
+    fun `unsafe get with missing bean component`() {
+        val ctx = Koin().build()
 
-        val serviceB = ctx.get<ServiceB>()
-        val serviceA = ctx.getOrNull<ServiceA>()
-        val serviceC = ctx.getOrNull<ServiceC>()
+        try {
+            assertNull(ctx.get<ServiceA>())
+        } catch(e: NoBeanDefFoundException) {
+            assertNotNull(e)
+        }
 
-        assertEquals(3, ctx.beanRegistry.definitions.size)
-        assertEquals(3, ctx.beanRegistry.instanceFactory.instances.size)
-
-        assertNotNull(serviceB)
-        assertNotNull(serviceA)
-        assertNotNull(serviceC)
+        assertEquals(0, ctx.beanRegistry.definitions.size)
+        assertEquals(0, ctx.beanRegistry.instanceFactory.instances.size)
     }
 
+    @Test
+    fun `missing bean component - lazy linking`() {
+        val ctx = Koin().build(SampleModuleC::class)
+
+        assertNull(ctx.getOrNull<ServiceA>())
+        assertNull(ctx.getOrNull<ServiceC>())
+        assertEquals(2, ctx.beanRegistry.definitions.size)
+        assertEquals(0, ctx.beanRegistry.instanceFactory.instances.size)
+    }
 
     @Test
-    fun `functional declaration with mock`() {
-        //onLoad only ServiceB
-        val ctx = Koin().build(SampleModuleB::class)
+    fun `functional declaration and inject mock`() {
+        val ctx = Koin().build()
 
         val serviceB: ServiceB = Mockito.mock(ServiceB::class.java)
 
@@ -62,23 +106,16 @@ class KoinTest {
             println("done B Mock")
         }
 
-        val serviceB_fromDI = ctx.get<ServiceB>()
-        val serviceA = ctx.getOrNull<ServiceA>()!!
+        val serviceA = ctx.get<ServiceA>()
         serviceA.doSomethingWithB()
 
-        assertNotNull(serviceB_fromDI)
         assertNotNull(serviceA)
         Mockito.verify(serviceB).doSomething()
     }
 
     @Test
     fun `factory declaration with mock`() {
-        //onLoad only ServiceB
-        val ctx = Koin().build(SampleModuleB::class)
-
-        assertEquals(1, ctx.beanRegistry.definitions.size)
-        assertEquals(0, ctx.beanRegistry.instanceFactory.instances.size)
-
+        val ctx = Koin().build()
 
         val serviceB: ServiceB = Mockito.mock(ServiceB::class.java)
         Mockito.`when`(serviceB.doSomething()).then {
@@ -87,37 +124,6 @@ class KoinTest {
 
         ctx.provide { serviceB }
         ctx.factory({ ServiceA(ctx.get()) })
-
-        val serviceA_1 = ctx.get<ServiceA>()
-        serviceA_1.doSomethingWithB()
-
-        assertEquals(2, ctx.beanRegistry.definitions.size)
-        assertEquals(2, ctx.beanRegistry.instanceFactory.instances.size)
-
-        val serviceA_2 = ctx.get<ServiceA>()
-        serviceA_2.doSomethingWithB()
-
-        assertEquals(2, ctx.beanRegistry.definitions.size)
-        assertEquals(2, ctx.beanRegistry.instanceFactory.instances.size)
-
-        assertNotEquals(serviceA_1, serviceA_2)
-        Mockito.verify(serviceB, times(2)).doSomething()
-    }
-
-    @Test
-    fun `use of factory instances`() {
-        //onLoad only ServiceB
-        val ctx = Koin().build(SampleModuleB::class)
-
-        val serviceB: ServiceB = Mockito.mock(ServiceB::class.java)
-        Mockito.`when`(serviceB.doSomething()).then {
-            println("done B Mock")
-        }
-        ctx.provide { serviceB }
-        ctx.factory(ServiceA::class)
-
-        assertEquals(2, ctx.beanRegistry.definitions.size)
-        assertEquals(0, ctx.beanRegistry.instanceFactory.instances.size)
 
         val serviceA_1 = ctx.get<ServiceA>()
         serviceA_1.doSomethingWithB()
@@ -160,23 +166,14 @@ class KoinTest {
         assertEquals(0, ctx.beanRegistry.definitions.size)
         assertEquals(0, ctx.beanRegistry.instanceFactory.instances.size)
 
-        val origin = ServiceB()
-        ctx.provide { origin }
+        ctx.import(SampleModuleA::class)
 
         assertEquals(1, ctx.beanRegistry.definitions.size)
         assertEquals(0, ctx.beanRegistry.instanceFactory.instances.size)
 
-        ctx.import(SampleModuleA::class)
-
-        assertEquals(2, ctx.beanRegistry.definitions.size)
-        assertEquals(0, ctx.beanRegistry.instanceFactory.instances.size)
-
-        val serviceB = ctx.get<ServiceB>()
-        val serviceA = ctx.get<ServiceA>()
-
-        Assert.assertEquals(origin, serviceB)
-        Assert.assertNotNull(serviceB)
-        Assert.assertNotNull(serviceA)
+        ctx.provide { ServiceB() }
+        Assert.assertNotNull(ctx.get<ServiceA>())
+        Assert.assertNotNull(ctx.get<ServiceB>())
 
         assertEquals(2, ctx.beanRegistry.definitions.size)
         assertEquals(2, ctx.beanRegistry.instanceFactory.instances.size)