+++ src/test/kotlin/org/koin/test/koin/example/classes_for_inject.kt
@@ -1,7 +1,5 @@
 package org.koin.test.koin.example
 
-import org.koin.test.ServiceB
-import org.koin.test.ServiceC
 import javax.inject.Inject
 
 /**